--- tests/test_additional_coverage.py
@@ -1,339 +1,0 @@
-"""Additional Coverage Tests for Issue #491 Phase 2
-
-Quick coverage booster tests for modules with easy-to-test functionality.
-Target: Push coverage from 8.81% to 10% (need ~1.19% more).
-"""
-
-import tempfile
-from pathlib import Path
-
-import pytest
-
-# Test config modules with good coverage potential
-from kumihan_formatter.config import ConfigManager
-
-# Test ast_nodes module (currently 32% coverage)
-from kumihan_formatter.core.ast_nodes import Node, NodeBuilder
-
-# Test text_processor module (currently 36% coverage)
-from kumihan_formatter.core.utilities.text_processor import TextProcessor
-
-# Test simple_config module (currently 58% coverage)
-from kumihan_formatter.simple_config import SimpleConfig
-
-
-class TestSimpleConfigCoverage:
-    """Quick tests for SimpleConfig module"""
-
-    def test_simple_config_basic_functionality(self):
-        """Test SimpleConfig basic operations"""
-        config = SimpleConfig()
-        assert config is not None
-
-        # Test actual API methods
-        css_vars = config.get_css_variables()
-        assert isinstance(css_vars, dict)
-
-        theme_name = config.get_theme_name()
-        assert isinstance(theme_name, str)
-
-        # Test DEFAULT_CSS class variable
-        assert hasattr(config, "DEFAULT_CSS")
-        assert isinstance(config.DEFAULT_CSS, dict)
-
-    def test_simple_config_css_operations(self):
-        """Test SimpleConfig CSS-related operations"""
-        config = SimpleConfig()
-
-        # Test CSS variables access
-        css_vars = config.get_css_variables()
-        assert isinstance(css_vars, dict)
-
-        # Test default CSS constants
-        default_css = config.DEFAULT_CSS
-        assert "max_width" in default_css
-        assert "background_color" in default_css
-        assert "font_family" in default_css
-
-        # Test internal manager exists
-        assert hasattr(config, "_manager")
-        assert config._manager is not None
-
-
-class TestNodeBuilderAdvanced:
-    """Advanced tests for NodeBuilder"""
-
-    def test_node_builder_chaining(self):
-        """Test NodeBuilder method chaining"""
-        builder = NodeBuilder("section")
-        node = (
-            builder.content("Test content")
-            .css_class("test-class")
-            .id("test-id")
-            .attribute("data-test", "value")
-            .build()
-        )
-
-        assert node.type == "section"
-        assert node.content == "Test content"
-        assert node.get_attribute("class") == "test-class"
-        assert node.get_attribute("id") == "test-id"
-        assert node.get_attribute("data-test") == "value"
-
-    def test_node_multiple_classes(self):
-        """Test Node with multiple CSS classes"""
-        node = NodeBuilder("div").css_class("class1").css_class("class2").build()
-
-        # Should handle multiple class additions
-        assert "class" in node.attributes
-
-    def test_node_complex_content(self):
-        """Test Node with complex nested content"""
-        child1 = Node("span", "Child 1")
-        child2 = Node("span", "Child 2")
-
-        parent = NodeBuilder("div").content([child1, child2]).build()
-        assert isinstance(parent.content, list)
-        assert len(parent.content) == 2
-        assert parent.content[0] == child1
-        assert parent.content[1] == child2
-
-
-class TestTextProcessorCoverage:
-    """Coverage tests for TextProcessor"""
-
-    def test_text_processor_initialization(self):
-        """Test TextProcessor initialization"""
-        processor = TextProcessor()
-        assert processor is not None
-
-    def test_text_processor_basic_operations(self):
-        """Test basic text processing operations"""
-        processor = TextProcessor()
-
-        # Test different text processing methods
-        test_text = "  Hello World  "
-
-        # Test if basic processing methods exist
-        if hasattr(processor, "normalize_whitespace"):
-            result = processor.normalize_whitespace(test_text)
-            assert isinstance(result, str)
-
-        if hasattr(processor, "clean_text"):
-            result = processor.clean_text(test_text)
-            assert isinstance(result, str)
-
-        if hasattr(processor, "process_text"):
-            result = processor.process_text(test_text)
-            assert isinstance(result, str)
-
-
-class TestConfigManagerAdvanced:
-    """Advanced ConfigManager tests for coverage"""
-
-    def test_config_manager_validation_scenarios(self):
-        """Test ConfigManager validation in different scenarios"""
-        config = ConfigManager()
-
-        # Test validation with different config states
-        assert isinstance(config.validate(), bool)
-
-        # Test setting invalid values and validation
-        config.set("invalid_key", None)
-        validation_result = config.validate()
-        assert isinstance(validation_result, bool)
-
-    def test_config_manager_theme_operations(self):
-        """Test ConfigManager theme-related operations"""
-        config = ConfigManager()
-
-        # Test theme operations
-        theme_name = config.get_theme_name()
-        assert isinstance(theme_name, str)
-
-        # Test CSS variables
-        css_vars = config.get_css_variables()
-        assert isinstance(css_vars, dict)
-
-        # Test theme validation
-        is_valid = config.validate()
-        assert isinstance(is_valid, bool)
-
-    def test_config_manager_file_operations(self):
-        """Test ConfigManager file operations"""
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".toml", delete=False) as f:
-            f.write('[formatting]\noutput_format = "html"\n[theme]\nname = "default"\n')
-            temp_path = f.name
-
-        try:
-            # Test loading from file
-            config = ConfigManager(config_path=temp_path)
-            assert config is not None
-
-            # Test that config was loaded
-            config_dict = config.to_dict()
-            assert isinstance(config_dict, dict)
-
-        finally:
-            Path(temp_path).unlink(missing_ok=True)
-
-
-class TestNodeAttributeOperations:
-    """Comprehensive Node attribute tests"""
-
-    def test_node_attribute_edge_cases(self):
-        """Test Node attribute edge cases"""
-        node = Node("div", "content")
-
-        # Test setting None attribute
-        node.add_attribute("empty", None)
-        assert node.get_attribute("empty") is None
-
-        # Test setting empty string attribute
-        node.add_attribute("empty_string", "")
-        assert node.get_attribute("empty_string") == ""
-
-        # Test overwriting attribute
-        node.add_attribute("test", "value1")
-        node.add_attribute("test", "value2")
-        assert node.get_attribute("test") == "value2"
-
-    def test_node_has_attribute(self):
-        """Test Node has_attribute method"""
-        node = Node("div", "content")
-
-        # Test with no attributes
-        assert not node.has_attribute("nonexistent")
-
-        # Test with existing attribute
-        node.add_attribute("test", "value")
-        assert node.has_attribute("test")
-        assert not node.has_attribute("other")
-
-    def test_node_attribute_manipulation(self):
-        """Test Node attribute manipulation methods"""
-        node = Node("div", "content")
-
-        # Add and check attribute
-        node.add_attribute("test", "value")
-        assert node.has_attribute("test")
-        assert node.get_attribute("test") == "value"
-
-        # Test attribute overwriting
-        node.add_attribute("test", "new_value")
-        assert node.get_attribute("test") == "new_value"
-
-        # Test Node element type checking methods
-        assert node.is_block_element()  # div is a block element
-        assert not node.is_inline_element()
-        assert not node.is_heading()
-
-
-class TestConfigIntegrationScenarios:
-    """Integration scenarios for configuration system"""
-
-    def test_config_with_different_formats(self):
-        """Test config with different output formats"""
-        config = ConfigManager()
-
-        # Test setting different output formats
-        formats = ["html", "markdown", "txt"]
-        for fmt in formats:
-            config.set("output_format", fmt)
-            assert config.get("output_format") == fmt
-
-            # Validate after each change
-            validation = config.validate()
-            assert isinstance(validation, bool)
-
-    def test_config_css_variable_operations(self):
-        """Test CSS variable operations"""
-        config = ConfigManager()
-
-        # Get CSS variables
-        css_vars = config.get_css_variables()
-        assert isinstance(css_vars, dict)
-
-        # Test individual CSS variable access
-        for key in list(css_vars.keys())[:3]:  # Test first 3 keys
-            value = css_vars[key]
-            assert isinstance(value, str)
-
-    def test_config_nested_access(self):
-        """Test nested configuration access"""
-        config = ConfigManager()
-
-        # Test accessing nested config values
-        config_dict = config.to_dict()
-        assert isinstance(config_dict, dict)
-
-        # Test that we can iterate through config
-        for key, value in config_dict.items():
-            assert isinstance(key, str)
-            # Value can be any type
-            assert value is not None or value is None
-
-
-class TestQuickCoverageBoost:
-    """Quick tests to boost coverage efficiently"""
-
-    def test_multiple_node_types(self):
-        """Test creating multiple node types"""
-        node_types = ["div", "span", "p", "h1", "h2", "h3", "section", "article"]
-
-        for node_type in node_types:
-            node = Node(node_type, f"Content for {node_type}")
-            assert node.type == node_type
-            assert node_type in node.content
-
-            # Test type checking methods
-            if node_type in ["h1", "h2", "h3"]:
-                assert node.is_heading()
-                assert node.get_heading_level() in [1, 2, 3]
-            else:
-                assert not node.is_heading()
-
-            # Test block/inline detection
-            if node_type in ["div", "p", "h1", "h2", "h3"]:
-                assert node.is_block_element()
-            elif node_type in ["span"]:
-                assert node.is_inline_element()
-
-    def test_builder_with_various_attributes(self):
-        """Test builder with various HTML attributes"""
-        builder = NodeBuilder("div")
-
-        # Add various common HTML attributes using correct API
-        node = (
-            builder.id("test-id")
-            .css_class("test-class")
-            .attribute("data-value", "test-data")
-            .attribute("title", "Test Title")
-            .attribute("role", "button")
-            .style("color: red;")
-            .build()
-        )
-
-        # Verify all attributes were set
-        assert node.get_attribute("id") == "test-id"
-        assert node.get_attribute("class") == "test-class"
-        assert node.get_attribute("data-value") == "test-data"
-        assert node.get_attribute("title") == "Test Title"
-        assert node.get_attribute("role") == "button"
-        assert node.get_attribute("style") == "color: red;"
-
-    def test_config_error_handling(self):
-        """Test config error handling scenarios"""
-        config = ConfigManager()
-
-        # Test getting non-existent keys
-        result = config.get("completely_nonexistent_key")
-        assert result is None or isinstance(result, str)
-
-        # Test setting various data types
-        test_values = ["string", 123, True, None]
-        for i, value in enumerate(test_values):
-            key = f"test_key_{i}"
-            config.set(key, value)
-            retrieved = config.get(key)
-            assert retrieved == value
