+++ tests/test_processor_coverage.py
@@ -1,0 +1,88 @@
+"""Processor Coverage Tests
+
+Split from test_additional_coverage.py for 300-line limit compliance.
+Coverage booster tests for text processor and quick coverage modules.
+"""
+
+import pytest
+
+from kumihan_formatter.core.ast_nodes import Node, NodeBuilder
+from kumihan_formatter.core.utilities.text_processor import TextProcessor
+
+
+class TestTextProcessorCoverage:
+    """Coverage tests for TextProcessor"""
+
+    def test_text_processor_initialization(self):
+        """Test TextProcessor initialization"""
+        processor = TextProcessor()
+        assert processor is not None
+
+    def test_text_processor_basic_operations(self):
+        """Test basic text processing operations"""
+        processor = TextProcessor()
+
+        # Test different text processing methods
+        test_text = "  Hello World  "
+
+        # Test if basic processing methods exist
+        if hasattr(processor, "normalize_whitespace"):
+            result = processor.normalize_whitespace(test_text)
+            assert isinstance(result, str)
+
+        if hasattr(processor, "clean_text"):
+            result = processor.clean_text(test_text)
+            assert isinstance(result, str)
+
+        if hasattr(processor, "process_text"):
+            result = processor.process_text(test_text)
+            assert isinstance(result, str)
+
+
+class TestQuickCoverageBoost:
+    """Quick tests to boost coverage efficiently"""
+
+    def test_multiple_node_types(self):
+        """Test creating multiple node types"""
+        node_types = ["div", "span", "p", "h1", "h2", "h3", "section", "article"]
+
+        for node_type in node_types:
+            node = Node(node_type, f"Content for {node_type}")
+            assert node.type == node_type
+            assert node_type in node.content
+
+            # Test type checking methods
+            if node_type in ["h1", "h2", "h3"]:
+                assert node.is_heading()
+                assert node.get_heading_level() in [1, 2, 3]
+            else:
+                assert not node.is_heading()
+
+            # Test block/inline detection
+            if node_type in ["div", "p", "h1", "h2", "h3"]:
+                assert node.is_block_element()
+            elif node_type in ["span"]:
+                assert node.is_inline_element()
+
+    def test_builder_with_various_attributes(self):
+        """Test builder with various HTML attributes"""
+        builder = NodeBuilder("div")
+
+        # Add various common HTML attributes using correct API
+        node = (
+            builder.id("test-id")
+            .css_class("test-class")
+            .attribute("data-value", "test-data")
+            .attribute("title", "Test Title")
+            .attribute("role", "button")
+            .style("color: red;")
+            .build()
+        )
+
+        # Verify all attributes were set
+        assert node.get_attribute("id") == "test-id"
+        assert node.get_attribute("class") == "test-class"
+        assert node.get_attribute("data-value") == "test-data"
+        assert node.get_attribute("title") == "Test Title"
+        assert node.get_attribute("role") == "button"
+        assert node.get_attribute("style") == "color: red;"
